refactor: Reservation 조회 서비스 분리
