+++ src/main/java/roomescape/reservation/service/ReservationQueryService.java
@@ -1,0 +1,50 @@
+package roomescape.reservation.service;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.controller.response.MyReservationResponse;
+import roomescape.reservation.controller.response.ReservationResponse;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.waiting.service.ReservationWaitingQueryService;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ReservationQueryService {
+
+    private final ReservationRepository reservationRepository;
+    private final ReservationWaitingQueryService waitingQueryService;
+
+    @Transactional(readOnly = true)
+    public List<ReservationResponse> getFilteredReservations(Long themeId, Long memberId, LocalDate from,
+                                                             LocalDate to) {
+        if (themeId == null && memberId == null && from == null && to == null) {
+            return getAllReservations();
+        }
+
+        List<Reservation> reservations = reservationRepository.findFilteredReservations(themeId, memberId, from, to);
+
+        return ReservationResponse.from(reservations);
+    }
+
+    private List<ReservationResponse> getAllReservations() {
+        List<Reservation> reservations = reservationRepository.findAll();
+
+        return ReservationResponse.from(reservations);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MyReservationResponse> getReservations(Long memberId) {
+        List<MyReservationResponse> responses = new ArrayList<>();
+        List<Reservation> myReservations = reservationRepository.findByMemberId(memberId);
+
+        responses.addAll(MyReservationResponse.from(myReservations));
+        responses.addAll(waitingQueryService.getWaitingReservations(memberId));
+        return responses;
+    }
+}
