+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -2,10 +2,11 @@ package roomescape.time.service;
 
 import java.time.LocalTime;
 import java.util.List;
-import java.util.NoSuchElementException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.global.exception.InvalidArgumentException;
+import roomescape.global.exception.NotFoundException;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.time.controller.request.AvailableReservationTimeRequest;
 import roomescape.time.controller.request.ReservationTimeCreateRequest;
@@ -15,16 +16,11 @@ import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.ReservationTimeRepository;
 
 @Service
+@RequiredArgsConstructor
 public class ReservationTimeService implements ReservationTimeQueryService {
 
     private final ReservationTimeRepository reservationTimeRepository;
     private final ReservationRepository reservationRepository;
-
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository,
-                                  ReservationRepository reservationRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.reservationRepository = reservationRepository;
-    }
 
     @Transactional
     public ReservationTimeResponse open(ReservationTimeCreateRequest request) {
@@ -63,7 +59,7 @@ public class ReservationTimeService implements ReservationTimeQueryService {
     @Transactional(readOnly = true)
     public ReservationTime getReservationTime(Long id) {
         return reservationTimeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("예약 시간을 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundException("예약 시간을 찾을 수 없습니다."));
     }
 
     @Transactional(readOnly = true)
