+++ src/main/java/roomescape/waiting/service/ReservationWaitingQueryService.java
@@ -1,0 +1,8 @@
+package roomescape.waiting.service;
+
+import java.util.List;
+import roomescape.reservation.controller.response.MyReservationResponse;
+
+public interface ReservationWaitingQueryService {
+    List<MyReservationResponse> getWaitingReservations(Long memberId);
+}
